Changed the trackbar to be a global component
