+++ src/App.tsx
@@ -1,18 +1,20 @@
 import "./App.css";
 
-import { FC, ReactElement, useEffect, useRef, useState } from "react";
+import { FC, ReactElement, useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 
 import { authSelectors } from "./containers/auth/selectors";
-import './styles/main.scss'
-import './styles/components.scss'
-import { useGetPlaylistsQuery, useGetPlaylistTracksQuery, useGetSearchTrackResultQuery, useGetUserQuery } from "./api/apiSlice";
+import { useGetPlaylistsQuery, useGetUserQuery } from "./api/apiSlice";
 import { Header } from "./components/Header";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Home } from "./page/Home";
 import { UserProfile } from "./page/User";
 import { HomePageSkeleton } from "./components/ui/Skeleton";
 import { SpotifyTrackItem } from "./types";
+import { TrackBar } from "./components/homepage/TrackBar";
+
+import './styles/main.scss'
+import './styles/components.scss'
 
 const App: FC = (): ReactElement => {
   const [theme, setTheme] = useState("dark")
@@ -23,8 +25,6 @@ const App: FC = (): ReactElement => {
   const [searchResults, setSearchResults] = useState<SpotifyTrackItem[] | null>(null);
   const [trackbarVisible, setTrackbarVisible] = useState(false);
 
-
-  // TODO: You can access user data and now fetch user's playlists
   const { data: playlists } = useGetPlaylistsQuery()
   const { data: user } = useGetUserQuery(undefined, {
     skip: !accessToken
@@ -52,6 +52,13 @@ const App: FC = (): ReactElement => {
     setTrackbarVisible(true)
   }
 
+  const hideTrackPanel = () => {
+    if (trackbarVisible) {
+      setSelectedTrack(null)
+      setTrackbarVisible(false);
+    }
+  }
+
   return (
     <div className="App" >
       <BrowserRouter>
@@ -62,14 +69,15 @@ const App: FC = (): ReactElement => {
               <HomePageSkeleton /> :
               <Home
                 playlists={playlists}
-                selectedTrack={selectedTrack}
                 setSelectedTrack={setTrack}
-                trackbarVisible={trackbarVisible}
                 setTrackbarVisible={setTrackbarVisible}
               />}
           />
           <Route path="User" element={user && playlists && <UserProfile user={user} playlists={playlists} theme={theme} />} />
         </Routes>
+        <section className={`trackbar ${trackbarVisible ? "show" : ""}`}>
+          {selectedTrack && <TrackBar track={selectedTrack} hideTrack={hideTrackPanel} />}
+        </section>
       </BrowserRouter>
     </div>
   );
