+++ src/styles/components.scss
@@ -75,6 +75,13 @@ select {
             width: 55px;
             cursor: pointer;
         }
+
+
+    }
+
+    @media screen and (max-width: 500px) {
+        display: none;
+
     }
 
     &.large {
