+++ src/styles/css/components.css
@@ -181,6 +181,11 @@ select {
   width: 55px;
   cursor: pointer;
 }
+@media screen and (max-width: 500px) {
+  .avatar {
+    display: none;
+  }
+}
 .avatar.large {
   width: 300px;
   height: 300px;
