+++ src/styles/main.scss
@@ -51,8 +51,6 @@ header {
 section.sidebar,
 section.trackbar {
     @include Flex(column, flex-start, flex-start, 15px);
-    width: 0;
-    height: 100%;
     padding: 0 20px;
     overflow-x: hidden;
     overflow-y: auto;
@@ -61,29 +59,25 @@ section.trackbar {
 }
 
 section.trackbar {
-    visibility: hidden;
+    visibility: visible;
+    position: fixed;
+    top: 50%;
+    bottom: 50%;
+    right: 0;
+    width: 400px;
+    height: 80%;
+    transform: translate(120%, -50%);
     border-radius: 20px;
+    margin-right: 50px;
+    z-index: 1000;
+    box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
+    background-color: var(--background-color);
 
-    @media screen and (max-width: 768px) {
-        position: fixed;
-        z-index: 1000;
-        background-color: var(--background-color);
-
-        left: 0;
-        right: 0;
-        bottom: 0;
-        width: 60vw;
-        height: 50%;
-        margin: 10px auto;
-    }
 
     &.show {
         visibility: visible;
-        width: 600px;
+        transform: translate(0, -50%);
 
-        @media screen and (max-width: 768px) {
-            width: 70%;
-        }
     }
 
 
@@ -91,6 +85,7 @@ section.trackbar {
 
 section.sidebar {
     width: 500px;
+    height: 100%;
 
     div {
         padding-bottom: 10px;
@@ -143,10 +138,8 @@ section.main {
 
 html[data-theme="dark"] .bg {
     position: absolute;
-    top: 100px;
-    left: 0;
-    width: 100%;
-    height: 30vh;
+    width: 100vw;
+    height: 40vh;
     z-index: -1;
     opacity: .7;
     overflow-y: hidden;
@@ -157,7 +150,12 @@ html[data-theme="dark"] .bg {
 }
 
 html[data-theme="light"] .bg {
-    display: none;
+    position: absolute;
+    width: 100vw;
+    height: calc(100% + 100px);
+    z-index: -1;
+    opacity: 0;
+    background: var(--track-bg);
 }
 
 .profile-container {
@@ -165,6 +163,7 @@ html[data-theme="light"] .bg {
     width: 100%;
     max-width: 1200px;
     margin: 100px auto;
+    position: relative
 }
 
 .name-container {
@@ -187,11 +186,19 @@ html[data-theme="light"] .bg {
     .playlist-toggle-button {
         @include Flex(row, flex-start, center, 10px);
         @include FontProperties(24px, 400, 1.2);
+        color: var(--text-color);
         padding: 10px 20px;
         background: none;
         border: none;
         border-radius: 0;
         border-bottom: 1px solid transparent;
+
+        &:hover {
+            color: var(--primary-color);
+            cursor: pointer;
+            border-bottom: 1px solid var(--text-color);
+        }
+
 
         &.active-toggle {
             border-color: var(--text-color);
@@ -257,7 +264,7 @@ html[data-theme="light"] .bg {
     }
 
     .playlist-list-container {
-        --image-size: 200px;
+        --image-size: 150px;
     }
 
     .playlist-toggle-container {
